Try using 'go test' subprocesses instead.
In the parent commit, we tried using plugins.  Turns out there's one
big, big drawback to plugins.  Building one invokes 'gcc'.
Works; as long as you... have gcc, and don't mind that having that as
a dependency that you now have to track and potentially worry about.
I don't consider that minor.

So, okay.  What's using 'go test' child invocations look like?

Well, bad.  It looks... bad.

The output is a mess.  Every child process starts its output as if its
a top level thing.  Why this happens is perfectly understandable and
hard to be mad at, but nonetheless the result is maddeningly illegible.
Cleaning this up might be possible, but will require some additional
effort to be invested.

There's some more irritating problems than that, though.
Notice the entirely new package that's sprouted.

Not being able to use closures in the test assertions is unfortunate.

Having to put the test assertions in *entirely different files* than
their setup... now that's really painful.

(And yeah, sure, we can move the setup out to another package, too.
It's not pleasing, though.  Then what of the test is actually in the
package it's testing?  The entrypoint?  So then, I'd end up editting
the test setup and assertions in one package (but running tests on
that package would actually no-op; syke!) and then have to switch to
the other package to actually run it?  There's no angle this can be
sliced that results in any sort of _nice_.)

Flags like '-v' aren't inherited now, either.  Hooray.
We could try to hack that in, I suppose.  Detect '-test.v' arg.
But it's just another thing that needs patching up to be shipshape.

Speaking of which, '-run' arg isn't even *remotely* manageable
without massive effort.  The thing can contain regexps, for goodness
sake -- how can we possibly semantically modify those for passdown?

I'm not sure where to go with this.  Depending on gcc in tests is
undesirable.  But all these other effects from trying to use 'go test'
this way seem even *more* undesirable.