Demo: codegen matching our Map3StrInt benchmark!
Results: mixed.

The good news:

The codegen works.

We were able to wire it to the standard benchmarks (!  great success).

It is flat out faster than any other implementation to date.

The not-so-good news:

It's not _as_ fast as I wanted >:(

The strategy of using a callback ("fcb") for transmitting 'finished'
signals from child assemblers to their parents causes an allocation.

That single source of allocations turns out to be one of the most
dominant things on the pprof of the benchmark.  (And it would
absolutely be even worse if 'N' was larger than '3' -- an alloc here
shifts us from an O(1) to O(n) on fields.)

So.  Good to know!  Having end to end benchmarks is VERY exciting.

And we're going to have to go back to the drawing board on that
part involving a callback.