Comments speculating on how to improve templates.
The reuse level is currently *very* low.

Some of that is intentional -- I'm writing radically non-DRY things
on purpose in many cases, because I want to see what *actually*
repeats most often before I start trying to remove repetitions.

Some of it is not intentional; it's because I just have no idea
how to organize units of reuse when doing templating.

This latter group could improve.