Third checkpoint in progress towards map gen.
This seems to be getting pretty close: the only compile errors left
are about not finding "fromString" methods for key reification...
which is of course sensible, because that's not what those methods
have been called yet in the cases where they've shown up;
and they also haven't been placed on type-level styles yet.

So the next commit will be a little detour to touch over those methods
formerly known as "construct" and update them a bit to this pattern.