Refactor 'construct'->'fromString'.
Using a function name that has kind info in it feels better in many
small ways (it's a nice sanity check; if we add coercion helper
methods for more kinds, it'll sure matter; etc).

And, that method on String types belongs on the type, not just on the
representation side.

And, drop the 'NewT' method for String types; instead, have them
generate use 'pkg.Style.T.FromString' pattern.