Map gen now works.
Through compilation, at least.  No functional tests yet written.

Bunch of small fixes for maybe pointers.  (Somehow I had gotten their
embedding in the entry struct wrong, and that kinda splashed all over.)

The refactor of method names 'construct'->'fromString' also made a
bunch of the generated map code line up, since it was already written
to target that convention.

Then it's just the matter of adding the tests that compile.

And so far so good: combos of nullable values, not nullable,
and maybe with or without use of pointers all appear to compile!