Terribly important character.  Avoids alloc.
With use of pointer here: overall benchmarks if marshalling or suchlike
are perfectly reasonable and on par with the basicnode implementations
(the biggest cost is getting iterators, which is similar in each).

With the typo causing a copy and another heap escape: that extra alloc
causes a zonking 10% slowdown on selected marshalling benchmark.

(I'm using benchmarks on the 'realgen' package.  MapNStrMap3StrInt can
be exercised now -- isn't that neat?  Suppose I'll commit that next.)