First partial draft of union assemblers.
These are turning out to be *very* fun.

In the process of detailing how child assemblers work for unions,
it has become time to consider how separately-allocated child
assemblers are handled, and what their existence implies about
the rest of our logical rules how all the assemblers behave in concert.
Like so many things in this endeavor... the far-reachingness of
the implications of even seemingly very mild choices is... high.

I think it's quite likely that the current claims about union
assemblers not needing extra state for set/focus might turn out to be
redacted in upcoming commits, and for the strangest reason: we might
actually need that state to be persisted (outside of 'w') so that
the *reset* mechanism, of all things, can be efficient.
We'll see; still sort of tumbling that thought around in my mind.
There are various options.