CI: test with Go 1.14 and 1.15, drop 1.12
The common policy is to support the two latest versions of Go. Do that,
since our previous state meant a version of Go that's pretty old and
unused at this point.

Also simplify the travis yaml. First, GO111MODULE=on hasn't been needed
for a while, since it defaults to "on" if go.mod exists. We also don't
need redundant steps like "go mod download" or "go build"; "go test"
already does all of that for us. The "go fmt" check is also pointless,
as it only formats the code without actually complaining if it isn't
properly formatted.

Finally, update the Go version in go.mod to 1.14, since that's the
oldest version we test against and support at this point. This version
sets, for example, what language features we can use.

And, while at it, "go mod tidy".